Change function signature on Linux to match Windows/OSX.

+++ ENIGMAsystem/SHELL/Platforms/xlib/XLIBwindow.cpp
@@ -353,7 +353,7 @@ void window_mouse_set(int x,int y) {
 	XWarpPointer(disp,None,win,0,0,0,0,(int)x,(int)y);
 }
 
-void display_mouse_set(double x,double y) {
+void display_mouse_set(int x,int y) {
 	XWarpPointer(disp,None,DefaultRootWindow(disp),0,0,0,0,(int)x,(int)y);
 }
